week3/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
